chore: dokka をシングルモジュールに変更
